Add a button to link a property
