--- frontend/src/lib/components/inputs/BooleanInput.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	import { Checkbox } from 'flowbite-svelte'
-
-	export let value: any
-	export let disabled: boolean = false
-</script>
-
-<Checkbox bind:checked={value} {disabled}>{value === undefined ? 'Not set' : ''}</Checkbox>
